% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\title{Consult available data}
\usage{
available(var_id, ag)
}
\arguments{
\item{var_id}{Variable's ID.}

\item{ag}{Aggregation level}
}
\value{
A list
}
\description{
Check time periods during which data is available for the selected variable and regional level combination.
}
\examples{
\dontrun{
available(ag ="municipio",var_id = 4845)
available(ag ="meso",var_id = 4845)
}
}
