% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_complement.R
\docType{data}
\name{geo_complement}
\alias{geo_complement}
\title{geo_complement}
\format{
A data frame
\describe{
\item{CodIBGE}{IBGE Code}
\item{Município}{Municipalities}
\item{Mesorregiões}{Mesorregions}
\item{Microrregião}{Microrregions}
\item{\verb{Região de Saúde}}{State health department Regions}
\item{CRS}{Regional health coordination offices Regions}
\item{Coredes}{Regional Development Councils Regions}
\item{RF}{Planning Functional Regions}
}
}
\source{
State secretariats at the government of state of Rio Grande do Sul
}
\usage{
data(geo_complement)
}
\description{
Complementary regionalizations for Health and Education
}
\keyword{datasets}
