% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key as environment variable}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{The API key}
}
\value{
No return value. Called for side effects only (sets an environment variable).
}
\description{
This function allows users to set the API key.
}
\examples{
\dontrun{
set_api_key("your_api_key_here")
}


}
