% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{tree}
\usage{
tree()
}
\value{
A Json Tree
}
\description{
This function function plots the DataRS tree. It is useful for exploring the available datasets at DataRS.
}
\examples{
\dontrun{
deetree()
}

}
