% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\title{List, filter and sort variables' ID and name}
\usage{
vars(sort = "ASC")
}
\arguments{
\item{sort}{Optional. A character string. Either "ASC" to sort from A to Z or "DESC" to sort from Z to A.}

\item{filters}{Optional. A character string to filter from the list of variables.}
}
\value{
a tibble
}
\description{
List, filter and sort variables' ID and name
}
\examples{
\dontrun{
#Run this to see the full list of variables
vars()
vars(sort = "ASC")
}



}
