% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoagregall2.R
\name{geoagregall2}
\alias{geoagregall2}
\title{All Geographic Aggregation}
\usage{
geoagregall2(period = NULL, key)
}
\arguments{
\item{period}{The year to consult. It allows single string (ex:2010), vector(ex:c(2010,2022) or NULL)}

\item{key}{The key to access the DataVis API. Check www.datavis.dee.rs.gov.br to learn more.}
}
\value{
a tibble.
}
\description{
Check ID and name of all geographic aggregations in selected periods. Default is all periods.
}
\examples{
\dontrun{
geoagregall2()
geoagregall2(period = 2010)
}

}
